requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery~=5.3.0
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
drf-spectacular~=0.27.0 
social-auth-app-django~=5.4.0
social-auth-core~=4.5.0
Pillow>=10.0.0
django-versatileimagefield~=3.0